read only copy column table and export/import in SS

+++ ydb/core/tx/schemeshard/olap/operations/drop_table.cpp
@@ -287,7 +287,13 @@ private:
 
         if (isStandalone) {
             for (auto& shard : txState->Shards) {
-                context.OnComplete.DeleteShard(shard.Idx);
+                auto shardIdx = shard.Idx;
+                TShardInfo& shardInfo = context.SS->ShardInfos[shardIdx];
+                shardInfo.CountReferences = shardInfo.CountReferences == 0 ? 0 : shardInfo.CountReferences - 1;
+                context.SS->PersistShardCountReferences(db, shardIdx, shardInfo.CountReferences);
+                if (shardInfo.CountReferences == 0) {
+                    context.OnComplete.DeleteShard(shard.Idx);
+                }
             }
         }
